(refactor) moved server to twin module
